Piecewise constant expected accrual

eAccrual() computes the expected cumulative enrollment (accrual) given a set of piecewise constant enrollment rates and times.

eAccrual(
  x = 0:24,
  enrollRates = tibble::tibble(duration = c(3, 3, 18), rate = c(5, 10, 20))
)

Arguments

x
times at which enrollment is to be computed.

enrollRates
Piecewise constant enrollment rates expressed as a tibble with duration for each piecewise constant period and the rate of enrollment for that period.

Value

A vector with expected cumulative enrollment for the specified times.

Specification

The contents of this section are shown in PDF user manual only.

Examples

# Example: default
eAccrual()

#>  [1]   0   5  10  15  25  35  45  65  85 105 125 145 165 185 205 225 245 265 285
#> [20] 305 325 345 365 385 405