Multivariate Normal Distribution for Multivariate Maximum Statistics

Computes the distribution function of the multivariate normal distribution with maximum statistics for arbitrary limits and correlation matrices

Usage

pmvnorm_combo(
  lower,
  upper,
  group,
  mean,
  corr,
  algorithm = GenzBretz(maxpts = 1e+05, abseps = 1e-05),
  ...
)

Arguments

lower

the vector of lower limits of length n.

upper

the vector of upper limits of length n.

group

the vector of test statistics group.

mean

the mean vector of length n.

corr

the correlation matrix of dimension n.

algorithm

an object of class GenzBretz, Miwa or TVPACK specifying both the algorithm to be used as well as the associated hyper parameters.

...

additional parameters transfer to mvtnorm::pmvnorm

Details

Let $Z = Z_ij$ be a multivariate normal distribution. Here i is a group indicator and j is a within group statistics indicator. Let G_i = max(Z_ij) for all test within one group. This program are calculating the probability

$$Pr( lower < max(G) < upper )$$