This is the function to format group sequential design into interger sample size.

This is the function to format group sequential design into interger sample size.

Usage

# S3 method for gs_design
to_integer(x, sample_size = TRUE, ...)

Arguments

x

an object returned from gs_design_ahr, gs_design_wlr, or gs_design_combo

sample_size

TRUE or FALSE, indicting if to ceiling sample size to an even integer

...

additional arguments

Value

a list similar to the output of gs_design_ahr, gs_design_wlr, or gs_design_combo, but the sample size is an integer

Examples

library(dplyr)
gs_design_ahr() %>% to_integer()
#> $input
#> $input$enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.3
#> 2 All            2  26.6
#> 3 All           10  39.8
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $input$event
#> [1] 292
#> 
#> $input$analysis_time
#> NULL
#> 
#> $input$upper
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x55efe2899c20>
#> <environment: namespace:gsDesign2>
#> 
#> $input$upar
#> [1] 4.332634 2.339816 2.011793
#> 
#> $input$lower
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x55efe2899c20>
#> <environment: namespace:gsDesign2>
#> 
#> $input$lpar
#> [1] -1.281552      -Inf      -Inf
#> 
#> $input$test_lower
#> [1] TRUE
#> 
#> $input$test_upper
#> [1] TRUE
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$binding
#> [1] FALSE
#> 
#> $input$info_scale
#> [1] 2
#> 
#> $input$r
#> [1] 18
#> 
#> $input$tol
#> [1] 1e-06
#> 
#> 
#> $enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.3
#> 2 All            2  26.6
#> 3 All           10  39.8
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $bounds
#> # A tibble: 2 × 7
#>   Analysis Bound Probability Probability0     Z `~HR at bound` `Nominal p`
#>      <int> <chr>       <dbl>        <dbl> <dbl>          <dbl>       <dbl>
#> 1        1 Upper  0.141        0.00000737  4.33          0.600  0.00000737
#> 2        1 Lower  0.00000356   0.100      -1.28          1.16   0.9       
#> 
#> $analysis
#> # A tibble: 1 × 10
#>   Analysis  Time     N Events   AHR theta  info info0 info_frac info_frac0
#>      <int> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl>      <dbl>
#> 1        1  35.8   478   292. 0.684 0.380  71.7  73.0         1          1
#> 
#> attr(,"class")
#> [1] "non-binding" "ahr"         "gs_design"   "list"       
gs_design_wlr() %>% to_integer()
#> $input
#> $input$enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.1
#> 2 All            2  26.2
#> 3 All           10  39.3
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $input$event
#> [1] 289
#> 
#> $input$analysis_time
#> NULL
#> 
#> $input$binding
#> [1] FALSE
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$upper
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x55efe2899c20>
#> <environment: namespace:gsDesign2>
#> 
#> $input$upar
#> [1] 4.332634 2.339816 2.011793
#> 
#> $input$test_upper
#> [1] TRUE
#> 
#> $input$lower
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x55efe2899c20>
#> <environment: namespace:gsDesign2>
#> 
#> $input$lpar
#> [1] -1.281552      -Inf      -Inf
#> 
#> $input$test_lower
#> [1] TRUE
#> 
#> $input$weight
#> function (x, arm0, arm1, rho = 0, gamma = 0, tau = NULL) 
#> {
#>     n <- arm0$size + arm1$size
#>     p1 <- arm1$size/n
#>     p0 <- 1 - p1
#>     if (!is.null(tau)) {
#>         if (tau > 0) {
#>             x <- pmin(x, tau)
#>         }
#>     }
#>     esurv <- p0 * npsurvSS::psurv(x, arm0) + p1 * npsurvSS::psurv(x, 
#>         arm1)
#>     (1 - esurv)^rho * esurv^gamma
#> }
#> <bytecode: 0x55efe25879f8>
#> <environment: namespace:gsDesign2>
#> 
#> $input$info_scale
#> [1] 2
#> 
#> $input$approx
#> [1] "asymptotic"
#> 
#> $input$r
#> [1] 18
#> 
#> $input$tol
#> [1] 1e-06
#> 
#> 
#> $enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.1
#> 2 All            2  26.2
#> 3 All           10  39.3
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $bounds
#> # A tibble: 2 × 7
#>   Analysis Bound Probability Probability0     Z `~HR at bound` `Nominal p`
#>      <int> <chr>       <dbl>        <dbl> <dbl>          <dbl>       <dbl>
#> 1        1 Upper  0.143        0.00000737  4.33          0.601  0.00000737
#> 2        1 Lower  0.00000356   0.100      -1.28          1.16   0.9       
#> 
#> $analysis
#>   Analysis     Time   N Events       AHR     theta     info    info0 info_frac
#> 1        1 35.90413 472    289 0.6834515 0.3805995 71.65282 73.14688         1
#>   info_frac0
#> 1          1
#> 
#> attr(,"class")
#> [1] "non-binding" "wlr"         "gs_design"   "list"