gsDesign2: Group Sequential Design with Non-Constant Effect

The goal of 'gsDesign2' is to enable fixed or group sequential design under non-proportional hazards. To enable highly flexible enrollment, time-to-event and time-to-dropout assumptions, 'gsDesign2' offers piecewise constant enrollment, failure rates, and dropout rates for a stratified population. This package includes three methods for designs: average hazard ratio, weighted logrank tests in Yung and Liu (2019) doi:10.1111/biom.13196 , and MaxCombo tests. Substantial flexibility on top of what is in the 'gsDesign' package is intended for selecting boundaries.

See also

Author

Maintainer: Yujie Zhao yujie.zhao@merck.com

Authors:

Keaven Anderson keaven_anderson@merck.com

Yilong Zhang elong0527@gmail.com

Jianxiao Yang yangjx@ucla.edu

Nan Xiao nan.xiao1@merck.com

Other contributors:

Amin Shirazi ashirazist@gmail.com [contributor]

Ruixue Wang ruixue.wang@merck.com [contributor]

Yi Cui yi.cui@merck.com [contributor]

Ping Yang ping.yang1@merck.com [contributor]

Xin Tong Li xin.tong.li@merck.com [contributor]

Chenxiang Li chenxiang.li@merck.com [contributor]

Hiroaki Fukuda hiroaki.fukuda@merck.com [contributor]

Hongtao Zhang hongtao.zhang1@merck.com [contributor]

Yalin Zhu yalin.zhu@outlook.com [contributor]

John Blischak jdblischak@gmail.com [contributor]

Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]