Group sequential design power under risk difference

Group sequential design power under risk difference

Usage

gs_power_rd(
  p_c = tibble::tibble(Stratum = "All", Rate = 0.2),
  p_e = tibble::tibble(Stratum = "All", Rate = 0.15),
  N = tibble::tibble(Stratum = "All", N = c(40, 50, 60), Analysis = 1:3),
  rd0 = 0,
  ratio = 1,
  weight = c("un-stratified", "ss", "invar"),
  upper = gs_b,
  lower = gs_b,
  upar = list(par = gsDesign(k = length(N), test.type = 1, sfu = sfLDOF, sfupar =
    NULL)$upper$bound),
  lpar = list(par = c(qnorm(0.1), rep(-Inf, length(N) - 1))),
  info_scale = c(0, 1, 2),
  binding = FALSE,
  test_upper = TRUE,
  test_lower = TRUE,
  r = 18,
  tol = 1e-06
)

Arguments

p_c

rate at the control group

p_e

rate at the experimental group

rd0

treatment effect under super-superiority designs, the default is 0

ratio

experimental:control randomization ratio

upper

function to compute upper bound

lower

function to compare lower bound

upar

parameter to pass to upper

lpar

parameter to pass to lower

info_scale

binding

indicator of whether futility bound is binding; default of FALSE is recommended

test_upper

indicator of which analyses should include an upper (efficacy) bound; single value of TRUE (default) indicates all analyses; otherwise, a logical vector of the same length as info should indicate which analyses will have an efficacy bound

test_lower

indicator of which analyses should include a lower bound; single value of TRUE (default) indicates all analyses; single value FALSE indicated no lower bound; otherwise, a logical vector of the same length as info should indicate which analyses will have a lower bound

r

Integer, at least 2; default of 18 recommended by Jennison and Turnbull

tol

Tolerance parameter for boundary convergence (on Z-scale)

n

sample size

theta0

the standardized treatment effect under H0

Value

a tibble with columns Analysis, Bound, Z, Probability, theta, Time, AHR, Events

Examples

# --------------------- #
#      example 1        #
# --------------------- #
# un-stratified case with H0: rd0 = 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = "All",
                       Rate = .2),
  p_e = tibble::tibble(Stratum = "All",
                       Rate = .15),
  N = tibble::tibble(Stratum = "All",
                     N = c(20, 40, 60),
                     Analysis = 1:3),
  rd0 = 0,
  ratio = 1,
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2))
)
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000309128454189077, 0.0181812128501248, 0.0728412074630513, 0.057142981125861), Probability0 = c(0.000103505718147101, 0.00604848662480753, 0.0249830182388001, 0.1), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.629112460595102, 0.301109521038808, 0.195108392773747, -0.217297640197857), `Nominal p` = c(0.000103505718147101, 0.0060122073928069, 
#> 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(20, 40, 60), rd = c(0.05, 0.05, 0.05), rd0 = c(0, 0, 0), theta1 = c(0.294883912309794, 0.41702882811415, 0.510753918455249), theta0 = c(0, 0, 0), info = c(34.7826086956522, 69.5652173913043, 104.347826086957), info0 = c(34.6320346320346, 69.2640692640693, 103.896103896104), IF = c(0.333333333333333, 0.666666666666667, 1), IF0 = c(0.333333333333333, 0.666666666666667, 1))

# --------------------- #
#      example 2        #
# --------------------- #
# un-stratified case with H0: rd0 != 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = "All",
                       Rate = .2),
  p_e = tibble::tibble(Stratum = "All",
                       Rate = .15),
  N = tibble::tibble(Stratum = "All",
                     N = c(20, 40, 60),
                     Analysis = 1:3),
  rd0 = 0.005,
  ratio = 1,
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2))
)
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000309128454189077, 0.0181812128501248, 0.0728412074630513, 0.057142981125861), Probability0 = c(0.000116215929691249, 0.00679858869044458, 0.0280730654715689, 0.094932925962366), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.571201214535592, 0.275998568934928, 0.180597553496372, -0.190567876178071), `Nominal p` = c(0.000103505718147101, 
#> 0.0060122073928069, 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(20, 40, 60), rd = c(0.05, 0.05, 0.05), rd0 = c(0.005, 0.005, 0.005), theta1 = c(0.294883912309794, 0.41702882811415, 0.510753918455249), theta0 = c(0.029424494316825, 0.041612518928824, 0.0509647191437626), info = c(34.7826086956522, 69.5652173913043, 104.347826086957), info0 = c(34.6320346320346, 69.2640692640693, 103.896103896104), IF = c(0.333333333333333, 0.666666666666667, 1), IF0 = c(0.333333333333333, 0.666666666666667, 1))

# use spending function
gs_power_rd(
  p_c = tibble::tibble(Stratum = "All",
                       Rate = .2),
  p_e = tibble::tibble(Stratum = "All",
                       Rate = .15),
  N = tibble::tibble(Stratum = "All",
                     N = c(20, 40, 60),
                     Analysis = 1:3),
  rd0 = 0.005,
  ratio = 1,
  upper = gs_spending_bound,
  lower = gs_b,
  upar = list(sf = gsDesign::sfLDOF, total_spend = 0.025, param = NULL, timing = NULL),
  lpar = c(qnorm(.1), rep(-Inf, 2))
)
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000309128454189077, 0.0181809114164984, 0.0728406608330404, 0.057142981125861), Probability0 = c(0.000116215929691249, 0.00679846071884225, 0.0280728068339514, 0.094932925962366), Z = c(3.7103028732625, 2.51143380837984, 1.99305107738576, -1.2815515655446), `~Risk difference at bound` = c(0.571201214535592, 0.275999302106181, 0.1805978666862, -0.190567876178071), `Nominal p` = c(0.000103505718147101, 0.00601209165905841, 
#> 0.023127927155208, 0.9))
#> <rd> list(Analysis = 1:3, N = c(20, 40, 60), rd = c(0.05, 0.05, 0.05), rd0 = c(0.005, 0.005, 0.005), theta1 = c(0.294883912309794, 0.41702882811415, 0.510753918455249), theta0 = c(0.029424494316825, 0.041612518928824, 0.0509647191437626), info = c(34.7826086956522, 69.5652173913043, 104.347826086957), info0 = c(34.6320346320346, 69.2640692640693, 103.896103896104), IF = c(0.333333333333333, 0.666666666666667, 1), IF0 = c(0.333333333333333, 0.666666666666667, 1))

# --------------------- #
#      example 3        #
# --------------------- #
# stratified case under sample size weighting and H0: rd0 = 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.15, .2, .25)),
  p_e = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.1, .16, .19)),
  N = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3),
                     Analysis = rep(1:3, 3),
                     N = c(10, 20, 24, 18, 26, 30, 10, 20, 24)),
  rd0 = 0,
  ratio = 1,
  weight = "ss",
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2)))
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000437464827787746, 0.0237074382352033, 0.0795090157094822, 0.0470452989790815), Probability0 = c(0.000103505718147101, 0.0060370551421798, 0.0236920418909984, 0.1), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.455636816364528, 0.22784063133903, 0.165810214453055, -0.157378547055991), `Nominal p` = c(0.000103505718147101, 0.0060122073928069, 
#> 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(38, 66, 78), rd = c(0.0478947368421053, 0.0490909090909091, 0.0492307692307692), rd0 = c(0, 0, 0), theta1 = c(0.39001233732009, 0.52793101572752, 0.575697262838691), theta0 = c(0, 0, 0), info = c(66.3103175915211, 115.65174911323, 136.746143057504), info0 = c(66.0379947133018, 115.151579236764, 136.151243353564), IF = c(0.484915450694918, 0.845740483258803, 1), IF0 = c(0.485034092136871, 0.84576223029953, 1))

# --------------------- #
#      example 4        #
# --------------------- #
# stratified case under inverse variance weighting and H0: rd0 = 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.15, .2, .25)),
  p_e = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.1, .16, .19)),
  N = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3),
                     Analysis = rep(1:3, 3),
                     N = c(10, 20, 24, 18, 26, 30, 10, 20, 24)),
  rd0 = 0,
  ratio = 1,
  weight = "invar",
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2)))
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000442619131011453, 0.0239764907165267, 0.0802868846418275, 0.0466866994642883), Probability0 = c(0.000103505718147101, 0.00603739081397565, 0.0236948427857254, 0.1), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.449264949546824, 0.224669536163963, 0.163506919691737, -0.155177682012191), `Nominal p` = c(0.000103505718147101, 0.0060122073928069, 
#> 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(38, 66, 78), rd = c(0.0476559902604003, 0.0487467035142855, 0.0488731146435554), rd0 = c(0, 0, 0), theta1 = c(0.393572117677309, 0.532713706342048, 0.58089684220815), theta0 = c(0, 0, 0), info = c(68.2045975246413, 119.425546719792, 141.27261916863), info0 = c(67.9137123039562, 118.889755231219, 140.635185952259), IF = c(0.482787095801126, 0.845355224689644, 1), IF0 = c(0.482906975548855, 0.845377025857368, 1))

# --------------------- #
#      example 5        #
# --------------------- #
# stratified case under sample size weighting and H0: rd0 != 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.15, .2, .25)),
  p_e = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.1, .16, .19)),
  N = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3),
                     Analysis = rep(1:3, 3),
                     N = c(10, 20, 24, 18, 26, 30, 10, 20, 24)),
  rd0 = 0.02,
  ratio = 1,
  weight = "ss",
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2)))
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000437464827787746, 0.0237074382352033, 0.0795090157094822, 0.0470452989790815), Probability0 = c(0.000194249720652402, 0.010853106084956, 0.0400648305312733, 0.0743583638571124), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.285370893047473, 0.152698389680973, 0.116570784241889, -0.0716600329007417), `Nominal p` = c(0.000103505718147101, 
#> 0.0060122073928069, 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(38, 66, 78), rd = c(0.0478947368421053, 0.0490909090909091, 0.0492307692307692), rd0 = c(0.02, 0.02, 0.02), theta1 = c(0.39001233732009, 0.52793101572752, 0.575697262838691), theta0 = c(0.162527529622894, 0.21461740771593, 0.233367729863033), info = c(66.3103175915211, 115.65174911323, 136.746143057504), info0 = c(66.0379947133018, 115.151579236764, 136.151243353564), IF = c(0.484915450694918, 0.845740483258803, 1), IF0 = c(0.48503409213687, 0.845762230299529, 1))

# --------------------- #
#      example 6        #
# --------------------- #
# stratified case under inverse variance weighting and H0: rd0 != 0
gs_power_rd(
  p_c = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.15, .2, .25)),
  p_e = tibble::tibble(Stratum = c("S1", "S2", "S3"),
                       Rate = c(.1, .16, .19)),
  N = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3),
                     Analysis = rep(1:3, 3),
                     N = c(10, 20, 24, 18, 26, 30, 10, 20, 24)),
  rd0 = 0.03,
  ratio = 1,
  weight = "invar",
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2)))
#> <rd> list(Analysis = c(1, 2, 3, 1), Bound = c("Upper", "Upper", "Upper", "Lower"), Probability = c(0.000442616071601571, 0.0239762905513859, 0.0802862704722098, 0.0466867818971965), Probability0 = c(0.000267021191731168, 0.0145173239498558, 0.0517931749854139, 0.0631593864749684), Z = c(3.7103028732625, 2.51142701385654, 1.99304752265394, -1.2815515655446), `~Risk difference at bound` = c(0.196446555810209, 0.113237323550553, 0.0905773492251296, -0.0274912214620675), `Nominal p` = c(0.000103505718147101, 
#> 0.0060122073928069, 0.0231281217605481, 0.9))
#> <rd> list(Analysis = 1:3, N = c(38, 66, 78), rd = c(0.0476558461332765, 0.0487463720571671, 0.0488727587857381), rd0 = c(0.03, 0.03, 0.03), theta1 = c(0.393570995322202, 0.532710171833852, 0.580892707363436), theta0 = c(0.247229318063527, 0.327109720573614, 0.355769103356754), info = c(68.2046210702201, 119.425586051398, 141.272665286879), info0 = c(67.9137063446184, 118.88974365972, 140.635172114743), IF = c(0.482787104863624, 0.845355227133881, 1), IF0 = c(0.482906980689071, 0.845377026756286, 1))