Derive spending bound for MaxCombo group sequential boundary

Derive spending bound for MaxCombo group sequential boundary

Usage

gs_spending_combo(par = NULL, info = NULL, ...)

Arguments

par

a list with the following items: sf (class spending function), total_spend (total spend), param (any parameters needed by the spending function sf()), timing (a vector containing values at which spending function is to be evaluated or NULL if information-based spending is used), max_info (when timing is NULL, this can be input as positive number to be used with info for information fraction at each analysis)

info

statistical information at all analyses, at least up to analysis k

...

additional parameters transfered to par$sf.

Examples

# alpha-spending
par <- list(sf = gsDesign::sfLDOF, total_spend = 0.025)
gs_spending_combo(par, info = 1:3/3)
#> [1] 0.0001035057 0.0060483891 0.0250000000

# beta-spending
par <- list(sf = gsDesign::sfLDOF, total_spend = 0.2)
gs_spending_combo(par, info = 1:3/3)
#> [1] 0.02643829 0.11651432 0.20000000