Rounds sample size to an even number for equal design

Rounds sample size to an even number for equal design

Usage

to_integer(x, ...)

# S3 method for fixed_design
to_integer(x, sample_size = TRUE, ...)

# S3 method for gs_design
to_integer(x, sample_size = TRUE, ...)

Arguments

x

An object returned by fixed_design(), gs_design_ahr(), gs_design_wlr(), or gs_design_combo().

...

Additional parameters (not used).

sample_size

Logical, indicting if ceiling sample size to an even integer.

Value

A list similar to the output of fixed_design(), gs_design_ahr(), gs_design_wlr(), or gs_design_combo(), except the sample size is an integer.

Examples

library(dplyr)
library(tibble)
library(gsDesign2)

# Average hazard ratio
# \donttest{
x <- fixed_design("ahr",
  alpha = .025, power = .9,
  enroll_rate = tibble(stratum = "All", duration = 18, rate = 1),
  fail_rate = tibble(
    stratum = "All", duration = c(4, 100),
    fail_rate = log(2) / 12, hr = c(1, .6),
    dropout_rate = .001
  ),
  study_duration = 36
)
x %>% to_integer()
#> $input
#> $input$alpha
#> [1] 0.025
#> 
#> $input$power
#> [1] 0.9
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$study_duration
#> [1] 36
#> 
#> $input$weight
#> NULL
#> 
#> $input$rho
#> NULL
#> 
#> $input$gamma
#> NULL
#> 
#> $input$tau
#> NULL
#> 
#> $input$enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18     1
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $input$n
#> NULL
#> 
#> 
#> $enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18  25.8
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $analysis
#> # A tibble: 1 × 7
#>   Design     N Events  Time Bound alpha Power
#>   <chr>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 ahr      464   325.  35.9  1.96 0.025 0.900
#> 
#> $design
#> [1] "ahr"
#> 
#> attr(,"class")
#> [1] "fixed_design" "list"        

# FH
x <- fixed_design("fh",
  alpha = 0.025, power = 0.9,
  enroll_rate = tibble(stratum = "All", duration = 18, rate = 20),
  fail_rate = tibble(
    stratum = "All", duration = c(4, 100),
    fail_rate = log(2) / 12, hr = c(1, .6),
    dropout_rate = .001
  ),
  rho = 0.5, gamma = 0.5,
  study_duration = 36, ratio = 1
)
x %>% to_integer()
#> $input
#> $input$alpha
#> [1] 0.025
#> 
#> $input$power
#> [1] 0.9
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$study_duration
#> [1] 36
#> 
#> $input$weight
#> NULL
#> 
#> $input$rho
#> [1] 0.5
#> 
#> $input$gamma
#> [1] 0.5
#> 
#> $input$tau
#> NULL
#> 
#> $input$enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18    20
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $input$n
#> NULL
#> 
#> 
#> $enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18    21
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $analysis
#> # A tibble: 1 × 7
#>   Design     N Events  Time Bound alpha Power
#>   <chr>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 fh       378   264.  35.8  1.96 0.025 0.900
#> 
#> $design
#> [1] "fh"
#> 
#> $design_par
#> $design_par$rho
#> [1] 0.5
#> 
#> $design_par$gamma
#> [1] 0.5
#> 
#> 
#> attr(,"class")
#> [1] "fixed_design" "list"        

# MB
x <- fixed_design("mb",
  alpha = 0.025, power = 0.9,
  enroll_rate = tibble(stratum = "All", duration = 18, rate = 20),
  fail_rate = tibble(
    stratum = "All", duration = c(4, 100),
    fail_rate = log(2) / 12, hr = c(1, .6),
    dropout_rate = .001
  ),
  tau = 4,
  study_duration = 36, ratio = 1
)
x %>% to_integer()
#> $input
#> $input$alpha
#> [1] 0.025
#> 
#> $input$power
#> [1] 0.9
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$study_duration
#> [1] 36
#> 
#> $input$weight
#> NULL
#> 
#> $input$rho
#> NULL
#> 
#> $input$gamma
#> NULL
#> 
#> $input$tau
#> [1] 4
#> 
#> $input$enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18    20
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $input$n
#> NULL
#> 
#> 
#> $enroll_rate
#> # A tibble: 1 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All           18  23.9
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            4    0.0578   1          0.001
#> 2 All          100    0.0578   0.6        0.001
#> 
#> $analysis
#> # A tibble: 1 × 7
#>   Design     N Events  Time Bound alpha Power
#>   <chr>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 mb       430   302.  36.1  1.96 0.025 0.901
#> 
#> $design
#> [1] "mb"
#> 
#> $design_par
#> $design_par$tau
#> [1] 4
#> 
#> 
#> attr(,"class")
#> [1] "fixed_design" "list"        
# }
# \donttest{
gs_design_ahr() %>% to_integer()
#> $input
#> $input$enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 all            2  13.3
#> 2 all            2  26.6
#> 3 all           10  39.8
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 all            3    0.0770   0.9        0.001
#> 2 all          100    0.0385   0.6        0.001
#> 
#> $input$event
#> [1] 292
#> 
#> $input$analysis_time
#> NULL
#> 
#> $input$upper
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x5567c4415d50>
#> <environment: namespace:gsDesign2>
#> 
#> $input$upar
#> [1] 4.332634 2.339816 2.011793
#> 
#> $input$lower
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x5567c4415d50>
#> <environment: namespace:gsDesign2>
#> 
#> $input$lpar
#> [1] -1.281552      -Inf      -Inf
#> 
#> $input$test_lower
#> [1] TRUE
#> 
#> $input$test_upper
#> [1] TRUE
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$binding
#> [1] FALSE
#> 
#> $input$info_scale
#> [1] 2
#> 
#> $input$r
#> [1] 18
#> 
#> $input$tol
#> [1] 1e-06
#> 
#> 
#> $enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 all            2  13.3
#> 2 all            2  26.6
#> 3 all           10  39.8
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 all            3    0.0770   0.9        0.001
#> 2 all          100    0.0385   0.6        0.001
#> 
#> $bound
#> # A tibble: 2 × 7
#>   analysis bound probability probability0     z `~hr at bound` `nominal p`
#>      <int> <chr>       <dbl>        <dbl> <dbl>          <dbl>       <dbl>
#> 1        1 upper  0.141        0.00000737  4.33          0.600  0.00000737
#> 2        1 lower  0.00000356   0.100      -1.28          1.16   0.9       
#> 
#> $analysis
#> # A tibble: 1 × 10
#>   analysis  time     n event   ahr theta  info info0 info_frac info_frac0
#>      <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl>      <dbl>
#> 1        1  35.8   478  292. 0.684 0.380  71.7  73.0         1          1
#> 
#> attr(,"class")
#> [1] "non_binding" "ahr"         "gs_design"   "list"       
# }
# \donttest{
gs_design_wlr() %>% to_integer()
#> $input
#> $input$enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.1
#> 2 All            2  26.2
#> 3 All           10  39.3
#> 
#> $input$fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $input$event
#> [1] 289
#> 
#> $input$analysis_time
#> NULL
#> 
#> $input$binding
#> [1] FALSE
#> 
#> $input$ratio
#> [1] 1
#> 
#> $input$upper
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x5567c4415d50>
#> <environment: namespace:gsDesign2>
#> 
#> $input$upar
#> [1] 4.332634 2.339816 2.011793
#> 
#> $input$test_upper
#> [1] TRUE
#> 
#> $input$lower
#> function (par = NULL, k = NULL, ...) 
#> {
#>     if (is.null(k)) {
#>         return(par)
#>     }
#>     else {
#>         return(par[k])
#>     }
#> }
#> <bytecode: 0x5567c4415d50>
#> <environment: namespace:gsDesign2>
#> 
#> $input$lpar
#> [1] -1.281552      -Inf      -Inf
#> 
#> $input$test_lower
#> [1] TRUE
#> 
#> $input$weight
#> function (x, arm0, arm1, rho = 0, gamma = 0, tau = NULL) 
#> {
#>     n <- arm0$size + arm1$size
#>     p1 <- arm1$size/n
#>     p0 <- 1 - p1
#>     if (!is.null(tau)) {
#>         if (tau > 0) {
#>             x <- pmin(x, tau)
#>         }
#>     }
#>     esurv <- p0 * npsurvSS::psurv(x, arm0) + p1 * npsurvSS::psurv(x, 
#>         arm1)
#>     (1 - esurv)^rho * esurv^gamma
#> }
#> <bytecode: 0x5567c7ce62c0>
#> <environment: namespace:gsDesign2>
#> 
#> $input$info_scale
#> [1] 2
#> 
#> $input$approx
#> [1] "asymptotic"
#> 
#> $input$r
#> [1] 18
#> 
#> $input$tol
#> [1] 1e-06
#> 
#> 
#> $enroll_rate
#> # A tibble: 3 × 3
#>   stratum duration  rate
#>   <chr>      <dbl> <dbl>
#> 1 All            2  13.1
#> 2 All            2  26.2
#> 3 All           10  39.3
#> 
#> $fail_rate
#> # A tibble: 2 × 5
#>   stratum duration fail_rate    hr dropout_rate
#>   <chr>      <dbl>     <dbl> <dbl>        <dbl>
#> 1 All            3    0.0770   0.9        0.001
#> 2 All          100    0.0385   0.6        0.001
#> 
#> $bounds
#> # A tibble: 2 × 7
#>   analysis bound probability probability0     z `~hr at bound` `nominal p`
#>      <int> <chr>       <dbl>        <dbl> <dbl>          <dbl>       <dbl>
#> 1        1 upper  0.143        0.00000737  4.33          0.601  0.00000737
#> 2        1 lower  0.00000356   0.100      -1.28          1.16   0.9       
#> 
#> $analysis
#>   analysis     time   n event       ahr     theta     info    info0 info_frac
#> 1        1 35.90413 472   289 0.6834515 0.3805995 71.65282 73.14688         1
#>   info_frac0
#> 1          1
#> 
#> attr(,"class")
#> [1] "non_binding" "wlr"         "gs_design"   "list"       
# }